requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
scikit-image==0.23.2
scipy==1.13.0
pydantic==2.7.1
aiofiles==23.2.1
